fix: 차트 register 추상화
