+++ src/app/product/detail/art/[id]/page.tsx
@@ -11,6 +11,8 @@ import ArtProfit from '@/components/product/detail/art/ArtProfit';
 import ArtProductDetail from '@/components/product/detail/art/ArtProductDetail';
 import Link from 'next/link';
 import { useAddBookMark, useDeleteBookMark } from '@/factory/BookMark';
+import { useAuthStore } from '@/store/userAuth.store';
+import '../../../../plugin';
 const ArtDetailpage = (props: { params: { id: string } }) => {
   const [sort, setSort] = useState('profit');
   const url = props.params.id;
@@ -28,23 +30,27 @@ const ArtDetailpage = (props: { params: { id: string } }) => {
   const addmutation = useAddBookMark();
   const deletemutation = useDeleteBookMark();
 
+  const { isLoggedIn } = useAuthStore();
   const handleBookmarkClick = (
     productId: string | undefined,
     bookmark: boolean | undefined
   ) => {
     // 낙관적 업데이트를 위해 로컬 상태를 먼저 변경합니다.
+    if (isLoggedIn) {
+      setLocalData((prevData) =>
+        prevData ? { ...prevData, bookmark: !prevData.bookmark } : prevData
+      );
 
-    setLocalData((prevData) =>
-      prevData ? { ...prevData, bookmark: !prevData.bookmark } : prevData
-    );
-
-    if (!bookmark) {
-      addmutation.mutate({ productId, bookmark });
-    } else if (bookmark) {
-      deletemutation.mutate({ productId });
+      if (!bookmark) {
+        addmutation.mutate({ productId, bookmark });
+      } else if (bookmark) {
+        deletemutation.mutate({ productId });
+      }
+    } else {
+      alert('로그인이 필요한 서비스입니다.');
+      window.location.href = '/sign';
     }
   };
-
   return (
     <div className="overflow-x-hidden   desk:mx-3">
       <Container>
