+++ src/app/product/detail/building/[id]/page.tsx
@@ -14,6 +14,7 @@ import { CATEGORY } from '@/static/category';
 import Link from 'next/link';
 import { useAddBookMark, useDeleteBookMark } from '@/factory/BookMark';
 import { useAuthStore } from '@/store/userAuth.store';
+import '../../../../plugin';
 const BuildingDetailpage = (props: any) => {
   const [sort, setSort] = useState('profit');
   const url = props.params.id;
