+++ src/app/product/detail/content/[id]/page.tsx
@@ -12,6 +12,7 @@ import { getContentProductDetail } from '@/factory/ProductDetail/ContentProductD
 import Link from 'next/link';
 import { useAddBookMark, useDeleteBookMark } from '@/factory/BookMark';
 import { useAuthStore } from '@/store/userAuth.store';
+import '../../../../plugin';
 const ContentDetailpage = (props: { params: { id: string } }) => {
   const [sort, setSort] = useState('profit');
   const url = props.params.id;
