+++ src/app/product/detail/cow/[id]/page.tsx
@@ -12,6 +12,7 @@ import { CATEGORY } from '@/static/category';
 import Link from 'next/link';
 import { useAddBookMark, useDeleteBookMark } from '@/factory/BookMark';
 import { useAuthStore } from '@/store/userAuth.store';
+import '../../../../plugin';
 const CowDetailpage = (props: { params: { id: string } }) => {
   const [sort, setSort] = useState('profit');
   const url = props.params.id;
