+++ src/app/product/detail/music/[id]/page.tsx
@@ -12,6 +12,7 @@ import MusicProductDetail from '@/components/product/detail/music/MusicProductDe
 import Link from 'next/link';
 import { useAddBookMark, useDeleteBookMark } from '@/factory/BookMark';
 import { useAuthStore } from '@/store/userAuth.store';
+import '../../../../plugin';
 
 const MusicDetailpage = (props: any) => {
   const [sort, setSort] = useState('profit');
