+++ src/components/product/detail/music/chart/YoutubeSearchChart.tsx
@@ -20,16 +20,28 @@ import {
   IContentYoutubeViewCharts
 } from '@/types/MusicProductType';
 
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  Title,
-  Tooltip,
-  Legend,
-  ChartDataLabels
-);
+// const customTextPlugin = {
+//   id: 'customText',
+//   beforeDraw: (chart: any) => {
+//     const { datasets } = chart.data;
+//     const hasData = datasets.some((dataset: any) =>
+//       dataset.data.some((value: number) => value !== 0)
+//     );
+//     if (!hasData) {
+//       const ctx = chart.ctx;
+//       const width = chart.width;
+//       const height = chart.height;
+//       ctx.save();
+//       ctx.textAlign = 'center';
+//       ctx.textBaseline = 'middle';
+//       ctx.font = '16px Arial';
+//       ctx.fillText('준비중입니다', width / 2, height / 2);
+//       ctx.restore();
+//     }
+//   }
+// };
+
+// ChartJS.register(customTextPlugin);
 
 const YoutubeSearchChart = () => {
   const pathname = usePathname();
